Title: Renamed library to M2RL
Body: Renamed the library to M2RL: Multipurpose Robotics Reinforcement
Learning library.

Authored-by: Jose Fajardo <jmfajardod@unal.edu.co>
