+++ setup.py
@@ -3,7 +3,7 @@ from catkin_pkg.python_setup import generate_distutils_setup
 
 # fetch values from package.xml
 setup_args = generate_distutils_setup(
-    packages=['gym_gazebo_sb3'],
+    packages=['m2rl'],
     package_dir={'': 'src'},
 )
