+++ src/m2rl/envs/robot_BasicEnv.py
@@ -2,13 +2,13 @@
 
 import gym
 from gym import spaces
-from gym_gazebo_sb3.common import ros_gazebo
-from gym_gazebo_sb3.common import ros_controllers
-from gym_gazebo_sb3.common import ros_node
-from gym_gazebo_sb3.common import ros_launch
-from gym_gazebo_sb3.common import ros_params
-from gym_gazebo_sb3.common import ros_urdf
-from gym_gazebo_sb3.common import ros_spawn
+from m2rl.common import ros_gazebo
+from m2rl.common import ros_controllers
+from m2rl.common import ros_node
+from m2rl.common import ros_launch
+from m2rl.common import ros_params
+from m2rl.common import ros_urdf
+from m2rl.common import ros_spawn
 import rospy
 
 class RobotBasicEnv(gym.Env):
